Fix saved address coordinate precision handling

+++ src/main/java/com/foodify/server/modules/addresses/dto/CoordinatesDto.java
@@ -6,18 +6,20 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class CoordinatesDto {
     @NotNull
     @DecimalMin(value = "-90.0")
     @DecimalMax(value = "90.0")
-    private Double latitude;
+    private BigDecimal latitude;
 
     @NotNull
     @DecimalMin(value = "-180.0")
     @DecimalMax(value = "180.0")
-    private Double longitude;
+    private BigDecimal longitude;
 
     private String geohash;
 }
